Move LauncherShortcut to feature module
