+++ settings.gradle.kts
@@ -69,6 +69,7 @@ include(
 
 include(
     ":feature:widget:message-list",
+    ":feature:widget:shortcut",
     ":feature:widget:unread",
 )
